added 'unfloat' option. default is true - and has no effect on current plugin behavior. setting to false will cause the floated element to never snap back into place, but instead remain floated over its original position when normally it would snap back into place. (there is no noticeable difference to the user when this option is set to false).
this is useful for situations where you may change the position of the parent element, such as horizontal scrolling it, but you dont want the scrolled element to snap back into its original spot which is now scrolled elsewhere on the page. 

another usecase might be modifying the floated elements dimensions and not wanting it to reset back to original size after being popped back into place.